feat: create and join communities
Ahora se pueden crear comunidades, hay algunos retoques por hacer aun

Refs#2

+++ app/modules/community/services.py
@@ -1,7 +1,5 @@
 from app.modules.community.repositories import CommunityRepository
 from app.modules.community.repositories import CommunityUsersRepository
-from app.modules.community.models import CommunityUser
-from app.modules.community.models import Community
 from core.services.BaseService import BaseService
 
 
@@ -15,14 +13,18 @@ class CommunityUserService(BaseService):
     def get_by_user_id(self, user_id):
         return self.repository.get_by_user_id(user_id)
 
-    def create(self, code, user_id, community_id, is_admin=False):
+    def get_by_user_id_and_community(self, community_id, user_id):
+        return self.repository.get_community_user_by_user_id_and_community(user_id=user_id, community_id=community_id)
+
+    def create(self, user_id, community_id, is_admin=False):
         # Check if community exists
-        community = Community.query.filter_by(code=code).first()
+        community = self.repository.get_by_id(id=community_id)
         if not community:
             return None
 
         # Check if user is already in community
-        community_user = CommunityUser.query.filter_by(user_id=user_id, community_id=community.id).first()
+        community_user = self.repository.get_community_user_by_user_id_and_community(
+            user_id=user_id, community_id=community_id)
         if community_user:
             return None
 
@@ -32,31 +34,30 @@ class CommunityUserService(BaseService):
             community_id=community_id,
             is_admin=is_admin
         )
-        self.repository.save(community_user)
+        self.repository.create(community_user.user_id, community_user.community_id, community)
         return community_user
 
-    def delete(self, user_id, community_id):
-        community_user = CommunityUser.query.filter_by(user_id=user_id, community_id=community_id).first()
-        if not community_user:
-            return False
+    def delete(self, community_user):
         self.repository.delete(community_user)
         return True
 
     def make_admin(self, user_id, community_id):
-        community_user = CommunityUser.query.filter_by(user_id=user_id, community_id=community_id).first()
+        community_user = self.repository.get_community_user_by_user_id_and_community(
+            user_id=user_id, community_id=community_id)
         if not community_user:
             return False
         community_user.is_admin = True
-        self.repository.save(community_user)
+        self.repository.create(community_user)
         return True
 
     def remove_admin(self, user_id, community_id):
-        community_user = CommunityUser.query.filter_by(user_id=user_id, community_id=community_id).first()
+        community_user = self.repository.get_community_user_by_user_id_and_community(
+            user_id=user_id, community_id=community_id)
         if not community_user:
             return False
 
         community_user.is_admin = False
-        self.repository.save(community_user)
+        self.repository.create(community_user)
         return True
 
 
@@ -68,6 +69,9 @@ class CommunityService(BaseService):
     def get_community_by_id(self, community_id):
         return self.repository.get_by_id(community_id)
 
+    def get_community_by_code(self, code):
+        return self.repository.get_by_code(code=code)
+
     def get_communities_by_user_id(self, user_id):
         return self.repository.get_communities_by_user_id(user_id)
 
@@ -75,18 +79,21 @@ class CommunityService(BaseService):
         return self.repository.get_communities_by_dataset_id(dataset_id)
 
     def create_community(self, name, description, code, owner):
-        community = self.repository.model(
+        community = self.repository.create(
             name=name,
             description=description,
             code=code
         )
-        self.community_user_service.create(code, user_id=owner.id, community_id=community.id, is_admin=True)
-        self.repository.save(community)
+        self.community_user_service.create(user_id=owner, community_id=community.id, is_admin=True)
         return community
 
     def update_community(self, community_id, name=None, code=None, description=None):
         community = self.repository.get_by_id(community_id)
         if not community:
+            return None
+        community_user_admin = self.community_user_service.get_by_user_id_and_community(
+            user_id=community.owner, community_id=community_id)
+        if not community_user_admin.is_admin:
             return None
         if name:
             community.name = name
@@ -94,15 +101,17 @@ class CommunityService(BaseService):
             community.description = description
         if code:
             community.code = code
-        self.repository.save(community)
+        self.repository.create(community)
         return community
 
     def delete_community(self, community_id):
         community = self.repository.get_by_id(community_id)
-        if community:
+        community_user_admin = self.community_user_service.get_by_user_id_and_community(
+            user_id=community.owner, community_id=community_id)
+        if community and community_user_admin.is_admin:
             self.repository.delete(community)
+            community_users = self.repository.get_users_by_community(community_id=community_id)
+            for community_user in community_users:
+                self.community_user_service.delete(community_user)
             return True
-        community_users = CommunityUser.query.filter_by(community_id=community_id).all()
-        for user in community_users:
-            self.community_user_service.delete(user.user_id, community_id)
         return False
